+++ crates/southesk-macros/src/lib.rs
@@ -1,0 +1,469 @@
+// Copyright 2026 Thomas Axelsson
+// SPDX-License-Identifier: MIT
+
+use indexmap::IndexMap;
+use quote::{ToTokens, quote};
+use serde::{Deserialize, Deserializer};
+
+use proc_macro2::{Ident, Span, TokenStream};
+use std::{fmt::Write, path::PathBuf};
+use syn::{
+    parse::{Parse, ParseStream},
+    parse_macro_input,
+};
+
+#[derive(Debug, Deserialize)]
+#[serde(rename_all = "camelCase")]
+struct Tool {
+    name: String,
+    description: String,
+    input_schema: JsType,
+    output_schema: JsType,
+}
+
+#[derive(Debug, Deserialize)]
+#[serde(rename_all = "camelCase")]
+struct McpSchema {
+    tools: Vec<Tool>,
+}
+
+impl Parse for McpSchema {
+    fn parse(input: ParseStream) -> syn::Result<Self> {
+        let json_path_lit = input.parse::<syn::LitStr>()?;
+        let json_path = PathBuf::from(&json_path_lit.value());
+        let json = std::fs::File::open(&json_path).unwrap_or_else(|_| {
+            panic!(
+                "Cannot open {} in {}",
+                json_path.display(),
+                std::env::current_dir().unwrap().display()
+            )
+        });
+        let schema: McpSchema = serde_json::from_reader(json).map_err(|e| {
+            syn::Error::new(json_path_lit.span(), format!("Failed to parse JSON: {}", e))
+        })?;
+        Ok(Self {
+            tools: schema.tools,
+        })
+    }
+}
+
+impl ToTokens for McpSchema {
+    fn to_tokens(&self, tokens: &mut TokenStream) {
+        let mut support_types = TokenStream::new();
+        let mut client_impl = TokenStream::new();
+
+        for tool in &self.tools {
+            tokenize_tool(tool, &mut client_impl, &mut support_types);
+        }
+
+        tokens.extend(quote! {
+            use rust_decimal::Decimal;
+            use types::*;
+
+            /// # Low-level API
+            ///
+            /// The following methods provides a direct mapping to the API
+            /// provided by the MCP. They are less ergonomic than the high-level
+            /// methods. Each method maps directly to a Montrose MCP tool of the
+            /// same name.
+            impl Client<Connected> {
+                #client_impl
+            }
+
+            /// Montrose Low-level API types
+            pub mod types {
+                use rust_decimal::Decimal;
+
+                #support_types
+            }
+        });
+    }
+}
+
+#[derive(Debug, PartialEq, Eq)]
+enum HasRef {
+    Yes,
+    No,
+}
+
+fn tokenize_tool(tool: &Tool, client_impl: &mut TokenStream, mut support_types: &mut TokenStream) {
+    let Tool {
+        name: tool_name,
+        description,
+        input_schema,
+        output_schema,
+    } = tool;
+
+    let provide_short_func = input_schema.properties.len() <= 4;
+    let args_struct_name = snake_to_pascal_case(tool_name) + "Args";
+    let ret_struct_name = snake_to_pascal_case(tool_name) + "Return";
+    let func_name = format!("low_{tool_name}");
+
+    let (arg_struct_ident, input_has_ref, _is_nullable) = process_type(
+        Some(&mut support_types),
+        &args_struct_name,
+        input_schema,
+        &format!("Arguments for [`{func_name}`](crate::Client::{func_name})"),
+        true,
+    );
+
+    // _Some_ of the tools have a top-level result property. It is actually
+    // never in the response, so strip it away.
+    let output_schema = if output_schema.properties.len() == 1
+        && let Some(res) = output_schema.properties.get("result")
+    {
+        res
+    } else {
+        output_schema
+    };
+    // If the response is just one plain value, we can return that directly,
+    // without creating a wrapping struct. By collapsing nested single
+    // values, we also get rid of the user having to type "return.value"
+    // for those output schemas that wrap the resulting value into extra
+    // single-member objects.
+    let mut return_obj = output_schema;
+    let mut return_member = TokenStream::new();
+    let mut return_type_name = ret_struct_name.clone();
+    while return_obj.properties.len() == 1 {
+        let (name, obj) = return_obj.properties.iter().next().unwrap();
+        return_obj = obj;
+        let var_name = Ident::new(&camel_to_snake_case(name), Span::call_site());
+        return_member.extend(quote! { . #var_name });
+        return_type_name += &camel_to_pascal_case(name);
+    }
+    let (return_type, _, _) = process_type(None, &return_type_name, return_obj, "", false);
+
+    let (ret_struct_ident, output_has_ref, _is_nullable) = process_type(
+        Some(&mut support_types),
+        &ret_struct_name,
+        output_schema,
+        &format!("Return value for [`{func_name}`](crate::Client::{func_name})."),
+        false,
+    );
+    assert!(
+        output_has_ref == Some(HasRef::No),
+        "Must return owned values"
+    );
+
+    let func_comment = format!("Low-level API. {description}");
+
+    // The (long) function taking a struct is only visible if a (short)
+    // function taking args directly is not visible.
+    let (long_func_pub, long_func_name) = if provide_short_func {
+        (quote! {}, &format!("{func_name}_long"))
+    } else {
+        (quote! { pub }, &func_name)
+    };
+    let long_func_name_ident = Ident::new(long_func_name, Span::call_site());
+
+    // serde derive macros do not duplicate allow(unused_lifetimes) on all
+    // generated structs, neither when putting it on the struct nor the mod, so
+    // we need to only add lifetime markers when needed.
+    let arg_lifetime = if input_has_ref == Some(HasRef::Yes) {
+        quote! { <'_> }
+    } else {
+        quote! {}
+    };
+
+    let long_func_body = quote! {
+        let json_args = serde_json::to_value(args)
+            .map_err(|e| ClientCallError::InvalidArguments(format!("Failed to serialize arguments: {e}")))?
+            .as_object()
+            .ok_or_else(|| ClientCallError::InvalidArguments(format!("JSON argument is not an object")))?
+            .to_owned();
+
+        // Question mark is needed when return_member is not empty
+        #[allow(clippy::needless_question_mark)]
+        Ok(self.api_call::<#ret_struct_ident>(#tool_name, Some(json_args)).await? #return_member)
+    };
+
+    let long_func_sig = quote! {
+        async fn #long_func_name_ident(&self, args: #arg_struct_ident #arg_lifetime) ->
+            Result<#return_type, ClientCallError>
+    };
+    client_impl.extend(quote! {
+        #[doc = #func_comment]
+        #long_func_pub #long_func_sig {
+            #long_func_body
+        }
+    });
+
+    if provide_short_func {
+        let short_func_name_ident = Ident::new(&func_name, Span::call_site());
+        let mut arg_mappings = TokenStream::new();
+        let mut short_func_args = TokenStream::new();
+        let mut short_func_comment = func_comment.clone();
+        write!(short_func_comment, "\n\n").unwrap();
+
+        for (prop_name, prop_js_type) in &input_schema.properties {
+            let arg_name = camel_to_snake_case(prop_name);
+            let prop_ident = Ident::new(&arg_name, Span::call_site());
+            let (mut prop_type_quote, _has_ref, _is_nullable) = process_type(
+                None,
+                &camel_to_pascal_case(prop_name),
+                prop_js_type,
+                "",
+                true,
+            );
+            if !input_schema.required.contains(prop_name) {
+                prop_type_quote = quote! { Option<#prop_type_quote> };
+            }
+            short_func_args.extend(quote! { #prop_ident: #prop_type_quote, });
+            arg_mappings.extend(quote! { #prop_ident, });
+            write!(
+                short_func_comment,
+                "`{}`: {}\n\n",
+                arg_name, prop_js_type.description
+            )
+            .unwrap();
+        }
+
+        let func_lifetime = if input_has_ref == Some(HasRef::Yes) {
+            quote! { <'arg> }
+        } else {
+            quote! {}
+        };
+
+        let short_func_body = quote! {
+            let args = #arg_struct_ident {
+                #arg_mappings
+            };
+
+            self.#long_func_name_ident(args).await
+        };
+
+        let short_func_sig = quote! {
+            async fn #short_func_name_ident #func_lifetime (&self, #short_func_args) ->
+                Result<#return_type, ClientCallError>
+        };
+        client_impl.extend(quote! {
+            #[doc = #short_func_comment]
+            pub #short_func_sig {
+                #short_func_body
+            }
+        });
+    }
+}
+
+/// Generates a Rust type, and optionally its needed support types, from a schema object.
+///
+/// The returned token stream matches what would be written as a struct member type.
+/// Examples: `i64`, `&'arg str`, `CallToolArgs`
+///
+/// HasRef will only be correctly set if support types are generated.
+fn process_type(
+    support_types: Option<&mut TokenStream>,
+    type_name_hint: &str,
+    js_type: &JsType,
+    comment: &str,
+    allow_ref: bool,
+) -> (TokenStream, Option<HasRef>, bool) {
+    let mut has_ref = HasRef::No;
+    let has_ref_valid = support_types.is_some();
+
+    let is_nullable = js_type.r#type.1;
+    let ident = match js_type.r#type.0.as_ref() {
+        "object" => {
+            let struct_ident = Ident::new(type_name_hint, Span::call_site());
+            if let Some(support_types) = support_types {
+                let mut struct_members = TokenStream::new();
+                js_type
+                    .properties
+                    .iter()
+                    .for_each(|(prop_name, child_object)| {
+                        let prop_ident =
+                            Ident::new(&camel_to_snake_case(prop_name), Span::call_site());
+                        let (mut prop_type, child_has_ref, child_is_nullable) = process_type(
+                            Some(support_types),
+                            &format!("{}{}", type_name_hint, camel_to_pascal_case(prop_name)),
+                            child_object,
+                            "",
+                            allow_ref,
+                        );
+                        if child_has_ref == Some(HasRef::Yes) {
+                            has_ref = HasRef::Yes;
+                        }
+                        let mut prop_attrs = TokenStream::new();
+                        if child_is_nullable || !js_type.required.contains(prop_name) {
+                            prop_type = quote! { Option<#prop_type> };
+                            #[cfg(not(feature = "__dev-macros"))]
+                            prop_attrs.extend(quote! {
+                                #[serde(skip_serializing_if = "Option::is_none")]
+                            });
+                        }
+                        let child_comment = if child_object.description.is_empty() {
+                            "(The MCP does not provide any documentation for this field.)"
+                        } else {
+                            &child_object.description
+                        };
+                        struct_members.extend(quote! {
+                            #[doc = #child_comment]
+                            #prop_attrs
+                            pub #prop_ident: #prop_type,
+                        });
+                    });
+                let arg_ref = if has_ref == HasRef::Yes {
+                    quote! { <'arg> }
+                } else {
+                    quote! {}
+                };
+                let comment = if comment.is_empty() {
+                    "(The MCP does not provide any documentation for this struct.)"
+                } else {
+                    comment
+                };
+                support_types.extend(quote! {
+                    #[doc = #comment]
+                });
+                #[cfg(not(feature = "__dev-macros"))]
+                {
+                    let mut struct_derives = quote! { Debug, Clone, PartialEq, serde::Serialize };
+                    if !allow_ref {
+                        struct_derives.extend(quote! { , serde::Deserialize });
+                    }
+                    support_types.extend(quote! {
+                        #[derive(#struct_derives)]
+                        #[serde(rename_all = "camelCase")]
+                    });
+                }
+                support_types.extend(quote! {
+                    pub struct #struct_ident #arg_ref {
+                        #struct_members
+                    }
+                });
+            }
+            quote! { #struct_ident }
+        }
+        "array" => {
+            let items = js_type.items.as_ref().expect("Array must have \"items\"");
+            let (item_type_ident, _child_has_ref, _child_is_nullable) = process_type(
+                support_types,
+                &format!("{type_name_hint}Item"),
+                items,
+                "",
+                allow_ref,
+            );
+            if allow_ref {
+                has_ref = HasRef::Yes;
+                quote! { &'arg [#item_type_ident] }
+            } else {
+                quote! { Vec<#item_type_ident> }
+            }
+        }
+        "string" => {
+            if allow_ref {
+                has_ref = HasRef::Yes;
+                quote! { &'arg str }
+            } else {
+                quote! { String }
+            }
+        }
+        "number" => quote! { Decimal },
+        "integer" => quote! { i64 },
+        "boolean" => quote! { bool },
+        _ => panic!("Unsupported type \"{:?}\"", js_type.r#type),
+    };
+    (
+        ident,
+        if has_ref_valid { Some(has_ref) } else { None },
+        is_nullable,
+    )
+}
+
+fn snake_to_pascal_case(s: &str) -> String {
+    let mut result = String::new();
+    let mut capitalize_next = true;
+    for ch in s.chars() {
+        if ch == '_' {
+            capitalize_next = true;
+        } else if capitalize_next {
+            result.push(ch.to_ascii_uppercase());
+            capitalize_next = false;
+        } else {
+            result.push(ch);
+        }
+    }
+    result
+}
+
+fn camel_to_snake_case(s: &str) -> String {
+    let mut result = String::new();
+    for ch in s.chars() {
+        if ch.is_ascii_uppercase() {
+            result.push('_');
+            result.push(ch.to_ascii_lowercase());
+        } else {
+            result.push(ch);
+        }
+    }
+    result
+}
+
+fn camel_to_pascal_case(s: &str) -> String {
+    let mut result = s.to_string();
+    if let Some(first) = result.get_mut(0..1) {
+        first.make_ascii_uppercase();
+    }
+    result
+}
+
+#[derive(Debug, Deserialize)]
+struct JsType {
+    #[serde(deserialize_with = "type_value")]
+    r#type: (String, bool),
+    #[serde(default)]
+    description: String,
+    #[serde(default)]
+    properties: IndexMap<String, JsType>,
+    #[serde(default)]
+    items: Option<Box<JsType>>,
+    #[serde(default)]
+    required: Vec<String>,
+}
+
+/// Deserializes the type member. It assumes that the first type is the wanted one.
+///
+/// Returns whether the type is nullable.
+fn type_value<'de, D: Deserializer<'de>>(deserializer: D) -> Result<(String, bool), D::Error> {
+    struct Type;
+
+    impl<'de> serde::de::Visitor<'de> for Type {
+        type Value = (String, bool);
+
+        fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
+            formatter.write_str("string or list of strings")
+        }
+
+        fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
+        where
+            E: serde::de::Error,
+        {
+            Ok((v.to_owned(), false))
+        }
+
+        fn visit_seq<A>(self, seq: A) -> Result<Self::Value, A::Error>
+        where
+            A: serde::de::SeqAccess<'de>,
+        {
+            let v: Vec<String> =
+                Deserialize::deserialize(serde::de::value::SeqAccessDeserializer::new(seq))
+                    .expect("sequence of strings");
+            let nullable = v.iter().any(|s| s == "null");
+            Ok((
+                v.first()
+                    .expect("sequence of one ore more strings")
+                    .to_owned(),
+                nullable,
+            ))
+        }
+    }
+
+    deserializer.deserialize_any(Type)
+}
+
+/// Creates MCP types from the JSON file at the given path.
+#[proc_macro]
+pub fn mcp_schema(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+    let c = parse_macro_input!(input as McpSchema);
+    quote! { #c }.into()
+}